--- script/token.js
@@ -1,3 +1,0 @@
-
-const token = '84c86aad-d0d1-4ec9-91d9-06e6853e2cef';
-const userId = 'ea1bff16927c3a7237205010';
