+++ src/index.html
@@ -5,13 +5,13 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mesto</title>
-  <link rel="stylesheet" href="./pages/index.css">
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
 </head>
 <body>
   <div class="root">
 
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('../images/logo.svg').default%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -31,7 +31,7 @@
   
   <template id='popup_type_add'>
         <div class="popup__content">
-          <img src="./images/close.svg" alt="" class="popup__close">
+          <img src="<%=require('../images/close.svg').default%>" alt="" class="popup__close">
           <h3 class="popup__title">Новое место</h3>
           <form class="popup__form" name="add">
               <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Название">
@@ -45,7 +45,7 @@
   
   <template id='popup_type_edit'>
         <div class="popup__content">
-          <img src="./images/close.svg" alt="" class="popup__close">
+          <img src="<%=require('../images/close.svg').default%>" alt="" class="popup__close">
           <h3 class="popup__title">Редактировать профиль</h3>
           <form class="popup__form" name="edit">
               <input type="text" name="name" class="popup__input popup__input_type_username" placeholder="Имя" value = 'Jacques Cousteau'>
@@ -59,7 +59,7 @@
   
   <template id='popup_type_avatar'>
     <div class="popup__content">
-      <img src="./images/close.svg" alt="" class="popup__close">
+      <img src="<%=require('../images/close.svg').default%>" alt="" class="popup__close">
       <h3 class="popup__title">Обновить аватар</h3>
       <form class="popup__form" name="edit">
           <input type="text" name="name" class="popup__input popup__input_type_link" placeholder="Ссылка на аватар">
@@ -73,7 +73,7 @@
   <template id='popup_type_image'>
         <div class="popup__content popup__content_type_image">
           <img src="#" alt="" class="popup__image">  
-          <img src="./images/close.svg" alt="" class="popup__close">
+          <img src="<%=require('../images/close.svg').default%>" alt="" class="popup__close">
         </div>
   </template>
  
@@ -91,15 +91,7 @@
       </div>
   </template>
     
-
-  <script src="script/token.js"></script>
-  <script src="script/Card.js"></script>
-  <script src="script/Api.js"></script>
-  <script src="script/CardList.js"></script>
-  <script src="script/PopupV2.js"></script>
-  <script src="script/validation.js"></script>
-  <script src="script/scriptV2.js"></script>
-  <script src="script/index.js"></script>
+  <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
 
 
 </body>
